+++ lib/callable_tree/node/internal/builder.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module CallableTree
+  module Node
+    module Internal
+      class Builder
+        include Node::Builder
+
+        def build
+          super(node_type: Internal)
+        end
+
+        private
+
+        def validate(matcher:, caller:, terminater:)
+          # noop
+        end
+      end
+    end
+  end
+end